+++ NEWS.md
@@ -2,7 +2,18 @@
 
 -------------------------------------------------------
 
-# aopdata v0.2.0 (dev)
+
+# aopdata v0.2.1 (dev)
+
+**Minor changes**
+* Allow for user to download data few selected cities. Closes #29
+* Fix  progress bar even when showProgress = F. Closes #30
+* Fix  encoding issues using `base::iconv(city, to="ASCII//TRANSLIT")`. Closes #31
+
+
+-------------------------------------------------------
+
+# aopdata v0.2.0
 
 **Major changes**
 * New function read_population. Closes #21